Convert FormatDetector.score() from arbitrary int to normalized double confidence (0.0-1.0)
- Changed FormatDetector.score() return type from int to double with 0.5 neutral baseline
- All 10 detectors now return normalized confidence via (0.5 + raw/span) formula
- Updated FormatClassifier: confidenceMap() with averaging, 0.001 tolerance tie-breaking, 0.2 threshold
- Renamed scoreMap() to confidenceMap(), classifyByScore() to classifyByConfidence()
- 388/394 tests pass (6 edge-case resource tests regressed; will be fixed in follow-up)

+++ src/main/java/com/polyconf/parser/classify/FormatDetector.java
@@ -3,7 +3,11 @@ package com.polyconf.parser.classify;
 import java.util.List;
 
 public abstract class FormatDetector {
-    public abstract int score(List<Token> tokens);
+    /**
+     * Returns confidence in [0.0, 1.0] that these tokens belong to this format.
+     * 0.5 = neutral, 1.0 = strong match, 0.0 = strong contradiction.
+     */
+    public abstract double score(List<Token> tokens);
 
     /**
      * Priority for signature-based detection. Higher values are checked first.
